prevent defer_info and undefer_sub from operating on expired subs
Validate that the arguments to defer_info and undefer_sub refer to
actual live subs.  Use the weak refs we are storing to the deferred and
undeferred subs to make sure the original subs are still alive, and we
aren't returning data related to a reused memory address.

Also make sure we don't expire data related to unnamed subs.  Since the
user can capture the undeferred sub via undefer_sub, we can't track the
expiry without using a fieldhash.  For now, avoid introducing that
complexity, since the amount we leak should not be that great.