work around buggy B::perlstring in perl 5.10.0
In perl 5.10.0, using B::perlstring on a utf8 flagged string will not
encode backslashes properly.  Usually it seem to just encode them as a
single backslash, although sometimes it seems it might read past the end
of strings.

Add a fallback escaping routine written manually that is used in this
case.  It uses a search and replace regex with a replacement table and
manual encoding for larger codepoints.  It's certainly much slower than
B::perlstring, but that is an acceptable compromise given the limited
cases it is used.