try harder to find accurate floating point representation
Floating point numbers can't always be accurately represented as decimal
numbers, but usually they can be at some rounding level.  Try to find
the shortest floating point representation that is accurate, up to the
maximum precision available.  If none of those work, fall back to the
longest.  While it won't always be perfect, floating point operations
should never be expected to be exact.