scale floats by exponent of 2 when needed for accurate representation
Floating point numbers can't always be represented in decimal
accurately.  This is because internally, they are represented using a
base of 2 rather than 10.  For numbers like this, we should be able to
scale the number by an exponent of 2 and have it maintain its full
precision.  This should allow us to scale the number until it fits in a
range that can be represented in decimal.  The quotified output can then
include the multiplication by the exponent of 2, allowing it to
reproduce the exact original value.